'Double-commit'

+++ src/models/platform/sale_item/sale_item.js
@@ -19,15 +19,41 @@ export async function addSaleItem(
   sale_item_total
 ) {
   try {
+    const { data: productData, error: productError } = await supabase
+      .from("products")
+      .select("quantity")
+      .eq("id", product_id)
+      .single();
+
+    if (productError) {
+      throw productError;
+    }
+
+    if (productData.quantity < quantity) {
+      throw new Error("No hay suficiente cantidad en el inventario.");
+    }
+
     const { data, error } = await supabase.from("sale_items").insert({
       sale_id: sale_id,
       product_id: product_id,
       quantity: quantity,
       sale_item_total: sale_item_total,
     });
+
     if (error) {
       throw error;
     }
+
+    const newQuantity = productData.quantity - quantity;
+    const { error: updateProductError } = await supabase
+      .from("products")
+      .update({ quantity: newQuantity })
+      .eq("id", product_id);
+
+    if (updateProductError) {
+      throw updateProductError;
+    }
+
     return data;
   } catch (error) {
     throw error;
@@ -78,14 +104,46 @@ export async function editSaleItem(
 
 export async function deleteSaleItem(sale_item_id) {
   try {
+    const { data: saleItemData, error: saleItemError } = await supabase
+      .from("sale_items")
+      .select("product_id, quantity")
+      .eq("id", sale_item_id)
+      .single();
+
+    if (saleItemError) {
+      throw saleItemError;
+    }
+
     const { data, error } = await supabase
       .from("sale_items")
       .delete()
       .eq("id", sale_item_id)
       .single();
+
     if (error) {
       throw error;
     }
+
+    const { data: productData, error: productError } = await supabase
+      .from("products")
+      .select("quantity")
+      .eq("id", saleItemData.product_id)
+      .single();
+
+    if (productError) {
+      throw productError;
+    }
+
+    const newQuantity = productData.quantity + saleItemData.quantity;
+    const { error: updateProductError } = await supabase
+      .from("products")
+      .update({ quantity: newQuantity })
+      .eq("id", saleItemData.product_id);
+
+    if (updateProductError) {
+      throw updateProductError;
+    }
+
     return data;
   } catch (error) {
     throw error;
@@ -117,7 +175,7 @@ export async function changeSaleItemQuantity(
       .from("sale_items")
       .update({
         quantity: quantity,
-        sale_item_total: sale_item_total
+        sale_item_total: sale_item_total,
       })
       .eq("id", sale_item_id);
     if (error) {
@@ -129,7 +187,13 @@ export async function changeSaleItemQuantity(
   }
 }
 
-export async function increaseSaleItemQuantity(sale_item_id, product_price, current_quantity, product_id, product_quantity) {
+export async function increaseSaleItemQuantity(
+  sale_item_id,
+  product_price,
+  current_quantity,
+  product_id,
+  product_quantity
+) {
   try {
     const newQuantity = current_quantity + 1;
     const newSaleItemTotal = product_price * newQuantity;
@@ -149,7 +213,7 @@ export async function increaseSaleItemQuantity(sale_item_id, product_price, curr
     const { data: productData, error: productError } = await supabase
       .from("products")
       .update({
-        quantity: product_quantity - 1 
+        quantity: product_quantity - 1,
       })
       .eq("id", product_id);
 
@@ -157,14 +221,19 @@ export async function increaseSaleItemQuantity(sale_item_id, product_price, curr
       throw productError;
     }
 
-    return saleItemData; 
+    return saleItemData;
   } catch (error) {
     throw error;
   }
 }
 
-
-export async function decreaseSaleItemQuantity(sale_item_id, product_price, current_quantity, product_id, product_quantity) {
+export async function decreaseSaleItemQuantity(
+  sale_item_id,
+  product_price,
+  current_quantity,
+  product_id,
+  product_quantity
+) {
   try {
     const newQuantity = current_quantity - 1;
     const newSaleItemTotal = product_price * newQuantity;
@@ -184,7 +253,7 @@ export async function decreaseSaleItemQuantity(sale_item_id, product_price, curr
     const { data: productData, error: productError } = await supabase
       .from("products")
       .update({
-        quantity: product_quantity + 1 
+        quantity: product_quantity + 1,
       })
       .eq("id", product_id);
 
@@ -192,9 +261,8 @@ export async function decreaseSaleItemQuantity(sale_item_id, product_price, curr
       throw productError;
     }
 
-    return saleItemData; 
+    return saleItemData;
   } catch (error) {
     throw error;
   }
 }
-
